backend/app/requirements.txt
python-dotenv
fastapi
uvicorn
neo4j
pandas
requests
openpyxl
faster-whisper
ollama
openai
websockets
neontology
python-jose
PyAudio
ffmpeg-python
scipy
websocket-client
jinja2
langchain[all]
langchain-community
pytest-html